refactor: 폴더 구조 변경 및 함수 스타일 통일
