+++ src/view/movieModal.ts
@@ -1,13 +1,13 @@
-export function openMovieModal() {
+export const openMovieModal = () => {
   const movieModal = document.querySelector("#modal-background");
   if (!movieModal) return;
 
   movieModal.classList.add("active");
-}
+};
 
-export function closeMovieModal() {
+export const closeMovieModal = () => {
   const movieModal = document.querySelector("#modal-background");
   if (!movieModal) return;
 
   movieModal.classList.remove("active");
-}
+};
