+++ src/view/skeleton.ts
