+++ test/api.test.ts
@@ -1,10 +1,10 @@
 import { describe, it, expect, vi } from "vitest";
 
-import { getMoviePopular } from "../src/services/api";
+import { getPopularMovies } from "../src/services/api";
 import { moviesFixture } from "./fixtures";
 
 describe("api 함수 호출 테스트", () => {
-  it("getMoviePopular 를 호출하면 20개의 영화 목록을 가지고 온다", async () => {
+  it("getPopularMovies 를 호출하면 20개의 영화 목록을 가지고 온다", async () => {
     const mockData = {
       page: 1,
       results: [...moviesFixture],
@@ -21,7 +21,7 @@ describe("api 함수 호출 테스트", () => {
 
     vi.stubGlobal("fetch", fetchMock);
 
-    const data = await getMoviePopular({ page: 1 });
+    const data = await getPopularMovies({ page: 1 });
 
     expect(data.results.length).toEqual(20);
   });
